I forgot to use Dtos so needed to add that. So i changes all controllers and all Services and IServices. Also did some minor changes in some other parts of the code

+++ Services/CustomerService.cs
@@ -1,5 +1,6 @@
 ﻿using Labb1_ResturantBookingSystem.Data.Repos.IRepos;
 using Labb1_ResturantBookingSystem.Models;
+using Labb1_ResturantBookingSystem.Models.DTOs;
 
 namespace Labb1_ResturantBookingSystem.Services
 {
@@ -12,30 +13,49 @@ namespace Labb1_ResturantBookingSystem.Services
             _customerRepository = customerRepository;
         }
 
-        public async Task<IEnumerable<Customer>> GetAllCustomersAsync()
+        public async Task<IEnumerable<CustomerDto>> GetAllCustomersAsync()
         {
-            return await _customerRepository.GetAllCustomersAsync();
+            var customers = await _customerRepository.GetAllCustomersAsync();
+            return customers.Select(c => new CustomerDto
+            {
+                Id = c.Id,
+                Name = c.Name,
+                PhoneNumber = c.PhoneNumber
+            });
         }
 
-        public async Task<Customer> GetCustomerByIdAsync(int id)
+        public async Task<CustomerDto> GetCustomerByIdAsync(int id)
         {
-            return await _customerRepository.GetCustomerByIdAsync(id);
+            var customer = await _customerRepository.GetCustomerByIdAsync(id);
+            if (customer == null) return null;
+
+            return new CustomerDto
+            {
+                Id = customer.Id,
+                Name = customer.Name,
+                PhoneNumber = customer.PhoneNumber
+            };
         }
 
-        public async Task CreateCustomerAsync(Customer customer)
+        public async Task CreateCustomerAsync(CreateCustomerDto createDto)
         {
+            var customer = new Customer
+            {
+                Name = createDto.Name,
+                PhoneNumber = createDto.PhoneNumber
+            };
+
             await _customerRepository.AddCustomerAsync(customer);
             await _customerRepository.SaveChangesAsync();
         }
 
-        public async Task UpdateCustomerAsync(int id, Customer updatedCustomer)
+        public async Task UpdateCustomerAsync(int id, CustomerDto updatedDto)
         {
             var customer = await _customerRepository.GetCustomerByIdAsync(id);
-            if (customer == null)
-                throw new ArgumentException("Customer not found.");
+            if (customer == null) throw new ArgumentException("Customer not found.");
 
-            customer.Name = updatedCustomer.Name;            
-            customer.PhoneNumber = updatedCustomer.PhoneNumber;
+            customer.Name = updatedDto.Name;
+            customer.PhoneNumber = updatedDto.PhoneNumber;
 
             await _customerRepository.UpdateCustomerAsync(customer);
             await _customerRepository.SaveChangesAsync();
@@ -43,6 +63,9 @@ namespace Labb1_ResturantBookingSystem.Services
 
         public async Task DeleteCustomerAsync(int id)
         {
+            var customer = await _customerRepository.GetCustomerByIdAsync(id);
+            if (customer == null) throw new ArgumentException("Customer not found.");
+
             await _customerRepository.DeleteCustomerAsync(id);
             await _customerRepository.SaveChangesAsync();
         }
